the real input data fft is working
The real input to the fft is working. The output matches the results of
Matlab fft. The only problem is that the current version can only
output the real part of the results.

Dealing with the complex input and output is under development. It is
the problem of matlab mex file, and will be fixed in next update.